võta, stipendium, lõputööprotokollid

+++ hois_back/src/main/java/ee/hitsa/ois/domain/scholarship/ScholarshipApplicationFamily.java
@@ -1,5 +1,7 @@
 package ee.hitsa.ois.domain.scholarship;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
@@ -13,11 +15,11 @@ public class ScholarshipApplicationFamily extends BaseEntityWithId {
     @JoinColumn(name = "scholarship_application_id", nullable = false, updatable = false)
     private ScholarshipApplication scholarshipApplication;
     private String name;
-    private Long netSalary;
-    private Long pension;
-    private Long stateBenefit;
-    private Long otherIncome;
-    private Long unemployedBenefit;
+    private BigDecimal netSalary;
+    private BigDecimal pension;
+    private BigDecimal stateBenefit;
+    private BigDecimal otherIncome;
+    private BigDecimal unemployedBenefit;
 
     public ScholarshipApplication getScholarshipApplication() {
         return scholarshipApplication;
@@ -35,43 +37,43 @@ public class ScholarshipApplicationFamily extends BaseEntityWithId {
         this.name = name;
     }
 
-    public Long getNetSalary() {
+    public BigDecimal getNetSalary() {
         return netSalary;
     }
 
-    public void setNetSalary(Long netSalary) {
+    public void setNetSalary(BigDecimal netSalary) {
         this.netSalary = netSalary;
     }
 
-    public Long getPension() {
+    public BigDecimal getPension() {
         return pension;
     }
 
-    public void setPension(Long pension) {
+    public void setPension(BigDecimal pension) {
         this.pension = pension;
     }
 
-    public Long getStateBenefit() {
+    public BigDecimal getStateBenefit() {
         return stateBenefit;
     }
 
-    public void setStateBenefit(Long stateBenefit) {
+    public void setStateBenefit(BigDecimal stateBenefit) {
         this.stateBenefit = stateBenefit;
     }
 
-    public Long getOtherIncome() {
+    public BigDecimal getOtherIncome() {
         return otherIncome;
     }
 
-    public void setOtherIncome(Long otherIncome) {
+    public void setOtherIncome(BigDecimal otherIncome) {
         this.otherIncome = otherIncome;
     }
 
-    public Long getUnemployedBenefit() {
+    public BigDecimal getUnemployedBenefit() {
         return unemployedBenefit;
     }
 
-    public void setUnemployedBenefit(Long unemployedBenefit) {
+    public void setUnemployedBenefit(BigDecimal unemployedBenefit) {
         this.unemployedBenefit = unemployedBenefit;
     }
