+++ hois_back/src/main/java/ee/hitsa/ois/domain/scholarship/ScholarshipTerm.java
@@ -1,5 +1,6 @@
 package ee.hitsa.ois.domain.scholarship;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -33,14 +34,14 @@ public class ScholarshipTerm extends BaseEntityWithId {
     private LocalDate paymentStart;
     private LocalDate paymentEnd;
     private Long places;
-    private Long amountPaid;
-    private Long averageMark;
+    private BigDecimal amountPaid;
+    private BigDecimal averageMark;
     @ManyToOne(optional = true, fetch = FetchType.LAZY)
     private Classifier averageMarkPriority;
-    private Long lastPeriodMark;
+    private BigDecimal lastPeriodMark;
     @ManyToOne(optional = true, fetch = FetchType.LAZY)
     private Classifier lastPeriodMarkPriority;
-    private Long curriculumCompletion;
+    private BigDecimal curriculumCompletion;
     @ManyToOne(optional = true, fetch = FetchType.LAZY)
     private Classifier curriculumCompletionPriority;
     private Long maxAbsences;
@@ -146,19 +147,19 @@ public class ScholarshipTerm extends BaseEntityWithId {
         this.places = places;
     }
 
-    public Long getAmountPaid() {
+    public BigDecimal getAmountPaid() {
         return amountPaid;
     }
 
-    public void setAmountPaid(Long amountPaid) {
+    public void setAmountPaid(BigDecimal amountPaid) {
         this.amountPaid = amountPaid;
     }
 
-    public Long getAverageMark() {
+    public BigDecimal getAverageMark() {
         return averageMark;
     }
 
-    public void setAverageMark(Long averageMark) {
+    public void setAverageMark(BigDecimal averageMark) {
         this.averageMark = averageMark;
     }
 
@@ -170,19 +171,19 @@ public class ScholarshipTerm extends BaseEntityWithId {
         this.averageMarkPriority = averageMarkPriority;
     }
 
-    public Long getLastPeriodMark() {
+    public BigDecimal getLastPeriodMark() {
         return lastPeriodMark;
     }
 
-    public void setLastPeriodMark(Long lastPeriodMark) {
+    public void setLastPeriodMark(BigDecimal lastPeriodMark) {
         this.lastPeriodMark = lastPeriodMark;
     }
 
-    public Long getCurriculumCompletion() {
+    public BigDecimal getCurriculumCompletion() {
         return curriculumCompletion;
     }
 
-    public void setCurriculumCompletion(Long curriculumCompletion) {
+    public void setCurriculumCompletion(BigDecimal curriculumCompletion) {
         this.curriculumCompletion = curriculumCompletion;
     }
